+++ buession-dao/src/main/java/org/apache/ibatis/type/DefaultEnumTypeHandler.java
@@ -21,7 +21,7 @@
  * +------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										|
  * | Author: Yong.Teng <webmaster@buession.com> 													|
- * | Copyright @ 2013-2022 Buession.com Inc.														|
+ * | Copyright @ 2013-2023 Buession.com Inc.														|
  * +------------------------------------------------------------------------------------------------+
  */
 package org.apache.ibatis.type;
@@ -40,8 +40,10 @@ import java.sql.SQLException;
  * 		枚举类型
  *
  * @author Yong.Teng
+ * @see org.apache.ibatis.type.EnumTypeHandler
  * @since 1.3.2
  */
+@Deprecated
 public class DefaultEnumTypeHandler<E extends Enum<E>> extends AbstractEnumTypeHandler<E> {
 
 	private final static Logger logger = LoggerFactory.getLogger(DefaultEnumTypeHandler.class);