+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - [依赖库版本升级和安全漏洞修复](https://github.com/buession/buession-parent/releases/tag/v2.3.0)
 
 
+### 🔔 变化
+- **buession-dao：** 废弃 DefaultEnumTypeHandler 使用 mybatis 原生 EnumTypeHandler
+
+
 ---